When parsing 'get bucket' responses, cast the object size to quint64 prior to assigning them to the QVariant list.  This fixes the build on certain platforms (e.g. Linux).
Remove the TODO comment right above this line.  Object sizes are already
humanized before being displayed to the user.

+++ src/models/ds3_browser_model.cc
@@ -871,8 +871,7 @@ DS3BrowserModel::HandleGetBucketResponse()
 
 			objectData << owner;
 
-			// TODO Humanize the size
-			objectData << rawObject.size;
+			objectData << (quint64)rawObject.size;
 
 			objectData << OBJECT;
